Split mindmap diagram transaction applying process
